HoYo

Description

hoyo is a command-line utility that lets the user save directories as bookmarks (similar to in the browser) and easily cd to them.

Bookmarks

data Bookmark

Bookmark a directory for easy cd. A bookmark remembers the directory, the index, the creation time, and optionally a user-specified nickname for the bookmark.

Constructors

Bookmark

Fields

_bookmarkDirectory :: !FilePath

_bookmarkIndex :: !Int

_bookmarkCreationTime :: !ZonedTime

_bookmarkName :: !(Maybe String)

newtype Bookmarks

Wrapper for [Bookmark].

Constructors

Bookmarks

Fields

unBookmarks :: [Bookmark]

searchBookmarks :: BookmarkSearchTerm -> Bookmarks -> ([Bookmark], [Bookmark])

searchBookmarks searchTerm bookmarks partitions bookmarks into a list of Bookmarks that match the search term and a list of those that do not.

filterBookmarks :: Maybe String -> Maybe String -> Bookmark -> Bool

Given a bookmark name and a bookmark directory, test if a bookmark matches those filters.

module HoYo.Bookmark

Config

data Config

A representation of hoyo settings.

Constructors

Config

Fields

_failOnError :: !Bool

_displayCreationTime :: !Bool

_enableClearing :: !Bool

_enableReset :: !Bool

_backupBeforeClear :: !Bool

defaultConfig :: Config

The default config for hoyo.

setConfig :: MonadError String m => String -> String -> Config -> m Config

Try to set a key-value pair in the config.

module HoYo.Config

module HoYo.Env

CLI commands

data Command

The core data-type for the hoyo CLI. The Command is parsed from the command-line, then runCommand dispatches on the type.

runCommand :: Command -> HoYoMonad ()

Run a Command in the hoyo environment.

module HoYo.Command

Utility functions

runHoYo :: HoYoMonad a -> Env -> IO (Either String a)

Given a hoyo Env, run a monadic action in IO.

data HoYoMonad a

HoYoMonad is the main monad stack for the hoyo program. It's essentially a wrapper around ExceptT String (ReaderT Env IO): in other words, HoYoMonad a is equivalent to Env -> IO (Either String a)

modifyBookmarks :: ([Bookmark] -> [Bookmark]) -> HoYoMonad ()

Helper function whenever we need to modify the saved bookmarks.

modifyBookmarks f retrieves the current bookmarks, applies f, and saves them back to file.

modifyBookmarksM :: ([Bookmark] -> HoYoMonad [Bookmark]) -> HoYoMonad ()

Helper function twhenever we need to modify the saved bookmarks, and need access to the hoyo environment.

modifyBookmarks f retrieves the current bookmarks, applies f in the hoyo environment, and saves them back to file.

printStderr :: MonadIO m => String -> m ()

Print to stderr.

printStdout :: MonadIO m => String -> m ()

Print to stdout.

readInt :: MonadError String m => String -> m Int

Try to read an Int.

readBool :: MonadError String m => String -> m Bool

Try to read a Bool.

backupFile :: (MonadIO m, MonadError String m) => FilePath -> String -> m ()

Try to back-up a file. Used when the "backup_before_clear" option is set.

assert :: String -> HoYoMonad Bool -> HoYoMonad ()

Throw an error if a check fails.

assertVerbose :: String -> HoYoMonad Bool -> HoYoMonad Bool

Throw an error if a check fails AND the "fail_on_error" flag is set.