HoYo.Env

Description

The read-only hoyo environment.

Documentation

data Env

The main hoyo read-only environment. Contains the current saved bookmarks, the current hoyo configuration, and the file locations for each.

Constructors

Env

Fields

_bookmarks :: !Bookmarks

_bookmarksPath :: !TFilePath

_config :: !Config

_configPath :: !TFilePath

initEnv :: MonadIO m => TFilePath -> TFilePath -> m ()

Given a filepath for the bookmarks file and a filepath for the config file, initialize the respective TOMLs at those locations.

getEnv :: MonadIO m => TFilePath -> TFilePath -> m (Either Text Env)

Retrieve an Env from given bookmark- and config- file locations.

writeEnv :: MonadIO m => Env -> m ()

Write an Env to file.

readEnv :: MonadIO m => TFilePath -> TFilePath -> m (Either Text Env)

Read an Env from a file.