HoYo.Internal.Bookmark

Description

Internals used by the HoYo.Bookmark module.

Documentation

bookmarkCodec :: TomlCodec Bookmark

A TomlCodec for encoding and decoding Bookmarks.

defaultBookmarkCodec :: TomlCodec DefaultBookmark

A TomlCodec for encoding and decoding Bookmarks.

bookmarksCodec :: TomlCodec Bookmarks

A TomlCodec for encoding and decoding Bookmarks.

decodeBookmarks :: Text -> Either Text Bookmarks

Decode a Bookmark from a Text.

decodeBookmarksFile :: MonadIO m => TFilePath -> m (Either Text Bookmarks)

Decode a Bookmark from a file.

encodeBookmarks :: Bookmarks -> Text

Encode a Bookmark to a Text.

encodeBookmarksFile :: MonadIO m => TFilePath -> Bookmarks -> m ()

Encode a Bookmark to a file.

searchBookmarks :: BookmarkSearchTerm -> Bookmarks -> ([Bookmark], [Bookmark])

searchBookmarks searchTerm bookmarks partitions bookmarks into a list of Bookmarks that match the search term and a list of those that do not.

filterBookmarkByName :: Maybe Text -> Bookmark -> Bool

A predicate used by filterBookmarks - match on the bookmark name. Note that matching is case-sensitive.

filterBookmarkByDirInfix :: Maybe Text -> Bookmark -> Bool

A predicate used by filterBookmarks - match on the bookmark directory.

filterBookmarks :: Maybe Text -> Maybe Text -> Bookmark -> Bool

Given a bookmark name and a bookmark directory, test if a bookmark matches those filters.

bookmarksFromDefault :: MonadIO m => [DefaultBookmark] -> m Bookmarks

Convert a list of DefaultBookmarks to a Bookmarks, assiging indices and/ creation times on the fly.

getBookmarks :: HoYoMonad Bookmarks

Get the bookmarks from the currently used bookmark file.