HoYo.Internal.Parse

Description

Parse CLI arguments.

Documentation

globalOptions :: Parser GlobalOptions

Parse global options: options that can be set on the command line no matter what sub-command is being run.

For a complete list of the global options, see GlobalOptions or run hoyo --help.

overrideOptions :: Parser OverrideOptions

Parse override options: options that override config settings. This can be useful when you want to temporarily enable functionality for one CLI run, but don't want to change it in the config file.

parseOverride :: Mod FlagFields Bool -> Mod FlagFields Bool -> Parser MaybeOverride

Parse a single override option as a pair of switches. For example parseOverrideFailOnError is defined as

parseOverrideFailOnError :: Parser MaybeOverride
parseOverrideFailOnError = parseOverride
                              (long "fail" <> help "Fail on error")
                              (long "nofail" <> help "Disable fail on error")

and results in the pair of switches:

 --fail                   Fail on error
 --nofail                 Disable fail on error

parseOverrideFailOnError :: Parser MaybeOverride

Parse a MaybeOverride corresponding to the "fail_on_error" config option.

parseOverrideDisplayCreationTime :: Parser MaybeOverride

Parse a MaybeOverride corresponding to the "display_creation_time" config option.

parseOverrideEnableClearing :: Parser MaybeOverride

Parse a MaybeOverride corresponding to the "enable_clear" config option.

parseOverrideEnableReset :: Parser MaybeOverride

Parse a MaybeOverride corresponding to the "enable_reset" config option.

parseOverrideBackupBeforeClear :: Parser MaybeOverride

Parse a MaybeOverride corresponding to the "backup_before_clear" config option.

addCommand :: Parser Command

Parse options for the hoyo add command.

bookmarkSearchTerm :: ReadM BookmarkSearchTerm

Parse a BookmarkSearchTerm. First tries to interpret the search term as a number corresponding to a bookmark index; if that fails, treat the term as a name.

moveCommandMods :: Mod ArgumentFields BookmarkSearchTerm

Parse mods + bookmark completer used in moveCommand and moveCommandHidden.

moveCommand :: Parser Command

Parse options for the hoyo move command.

moveCommandHidden :: Parser Command

When hoyo is run without a command but with a search term (e.g. `hoyo docs`), it is interpreted as a move command - so the example above would be interpreted as `hoyo move docs`. We hide this from the CLI help message because it's quite ugly.

listCommand :: Parser Command

Parse options for the hoyo list command.

clearCommand :: Parser Command

Parse options for the hoyo clear command.

deleteCommand :: Parser Command

Parse options for the hoyo delete command.

refreshCommand :: Parser Command

Parse options for the hoyo refresh command.

configCommand :: Parser Command

Parse options for the hoyo config command.

configPrintCommand :: Parser ConfigCommand

Parse options for the hoyo config print sub-command.

configResetCommand :: Parser ConfigCommand

Parse options for the hoyo config reset sub-command.

configSetCommand :: Parser ConfigCommand

Parse options for the hoyo config set sub-command.

configAddDefaultCommand :: Parser ConfigCommand

Parse options for the hoyo config add-default sub-command.

noArgs :: CheckOptions -> CheckOptions

`hoyo check` should be considered equivalent to `hoyo check --config --bookmarks`

checkCommand :: Parser Command

Parse options for the hoyo check command.

defaultCommand :: Parser Command

If hoyo is run with no arguments, we run the "default command" if it's set.

parseCommand :: Parser Command

Parse a command and the arguments/options for that command from the command-line arguments.

parseOptions :: Parser Options

Parse an Options argument, which includes the command to run and any global options.

versionInfo :: String

Version information printed by `hoyo --version`.

versionOption :: Parser (a -> a)

Add a hidden --version flag to the end of a parser.

options :: ParserInfo Options

A ParserInfo object containing the necessary information for parsing CLI commands and arguments, and displaying useful help text.

splitArgs :: Text -> [String]

Split a string into arguments as they would be interpreted on the command line.

Adapted from this StackOverflow comment.