HoYo.Internal.Types

Description

Types used by all the main HoYo.* modules.

Documentation

type TFilePath = Text

A Text version of FilePath.

data Env

The main hoyo read-only environment. Contains the current saved bookmarks, the current hoyo configuration, and the file locations for each.

Constructors

Env

Fields

_bookmarks :: !Bookmarks

_bookmarksPath :: !TFilePath

_config :: !Config

_configPath :: !TFilePath

data Bookmark

Bookmark a directory for easy cd. A bookmark remembers the directory, the index, the creation time, and optionally a user-specified nickname for the bookmark.

Constructors

Bookmark

Fields

_bookmarkDirectory :: !TFilePath

_bookmarkIndex :: !Int

_bookmarkCreationTime :: !ZonedTime

_bookmarkName :: !(Maybe Text)

data DefaultBookmark

Default bookmarks to save at init. A default bookmark remembers the directory and optionally a user-specified nickname for the bookmark.

Constructors

DefaultBookmark

Fields

_defaultBookmarkDirectory :: !TFilePath

_defaultBookmarkName :: !(Maybe Text)

newtype Bookmarks

Wrapper for [Bookmark].

Constructors

Bookmarks

Fields

unBookmarks :: [Bookmark]

data BookmarkSearchTerm

Data-type for represting a bookmark search. You can either search by index or by name. Used by the delete and move commands.

Constructors

SearchIndex Int

SearchName Text

data Config

A representation of hoyo settings.

Constructors

Config

Fields

_failOnError :: !Bool

_displayCreationTime :: !Bool

_enableClearing :: !Bool

_enableReset :: !Bool

_backupBeforeClear :: !Bool

_defaultBookmarks :: ![DefaultBookmark]

_defaultCommand :: !(Maybe Text)

newtype HoYoMonad a

HoYoMonad is the main monad stack for the hoyo program. It's essentially a wrapper around ExceptT T.Text (ReaderT Env IO): in other words, HoYoMonad a is equivalent to Env -> IO (Either T.Text a)

Constructors

HoYoMonad

Fields

unHoYo :: ExceptT Text (ReaderT Env IO) a

data ExecResult

The result of executing a command. Currently only used meaningfully by runDefaultCommand.

Constructors

Done

ShowHelp

ReRun Text

bookmarkName :: Lens' Bookmark (Maybe Text)

bookmarkIndex :: Lens' Bookmark Int

bookmarkDirectory :: Lens' Bookmark TFilePath

bookmarkCreationTime :: Lens' Bookmark ZonedTime

failOnError :: Lens' Config Bool

enableReset :: Lens' Config Bool

enableClearing :: Lens' Config Bool

displayCreationTime :: Lens' Config Bool

defaultCommand :: Lens' Config (Maybe Text)

defaultBookmarks :: Lens' Config [DefaultBookmark]

backupBeforeClear :: Lens' Config Bool

configPath :: Lens' Env TFilePath

config :: Lens' Env Config

bookmarksPath :: Lens' Env TFilePath

bookmarks :: Lens' Env Bookmarks