Hoyo

Description

hoyo is a command-line utility that lets the user save directories as bookmarks (similar to in the browser) and easily cd to them.

Bookmarks

data Bookmark

Bookmark a directory for easy cd. A bookmark remembers the directory, the index, the creation time, and optionally a user-specified nickname for the bookmark.

Constructors

Bookmark

Fields

_bookmarkDirectory :: !FilePath

_bookmarkIndex :: !Int

_bookmarkCreationTime :: !ZonedTime

_bookmarkName :: !(Maybe Text)

newtype Bookmarks

Wrapper for [Bookmark].

Constructors

Bookmarks

Fields

unBookmarks :: [Bookmark]

searchBookmarks :: BookmarkSearchTerm -> Bookmarks -> ([Bookmark], [Bookmark])

searchBookmarks searchTerm bookmarks partitions bookmarks into a list of Bookmarks that match the search term and a list of those that do not.

filterBookmarks :: Maybe Text -> Maybe Text -> Bookmark -> Bool

Given a bookmark name and a bookmark directory, test if a bookmark matches those filters.

module Hoyo.Bookmark

Config

data Config

A representation of hoyo settings.

Constructors

Config

Fields

_failOnError :: !(ConfigValue 'TBool)

_displayCreationTime :: !(ConfigValue 'TBool)

_enableClearing :: !(ConfigValue 'TBool)

_enableReset :: !(ConfigValue 'TBool)

_backupBeforeClear :: !(ConfigValue 'TBool)

_defaultBookmarks :: !(ConfigValue ('TList 'TDefaultBookmark))

_defaultCommand :: !(ConfigValue ('TMaybe 'TCommand))

defaultConfig :: Config

The default config for hoyo.

setConfig :: MonadError HoyoException m => Text -> Text -> Config -> m Config

Try to set a key-value pair in the config.

module Hoyo.Config

module Hoyo.Env

CLI commands

data Command

The core data-type for the hoyo CLI. The Command is parsed from the command-line, then runCommand dispatches on the type.

runCommand :: Command -> HoyoMonad ()

Run a Command in the hoyo environment.

module Hoyo.Command

Utility functions

runHoyo :: HoyoMonad a -> Env -> IO (Either HoyoException a)

Given a hoyo Env, run a monadic action in IO.

withFiles :: GlobalOptions -> FilePath -> FilePath -> HoyoMonad a -> IO (Either HoyoException a)

withFiles globals bFp sFp hoyo gets the environment saved in the bookmark path (bFp) and the config path (sFp), applies the global options and overrides in globals, and runs hoyo, returning either the result or an error message.

getEnvAndRunHoyo :: GlobalOptions -> HoyoMonad a -> FilePath -> FilePath -> IO a

getEnvAndRunHoyo globals hoyo bFp sFp gets the environment saved in the bookmark path (bFp) and the config path (sFp), applies the global options and overrides in globals, and runs hoyo, either printing an error message or discarding the result.

getEnvAndRunCommand :: Options -> FilePath -> FilePath -> IO ()

getEnvAndRunHoyo opts bFp sFp gets the environment saved in the bookmark path (bFp) and the config path (sFp), and runs the command specified by opts.

data HoyoException

A custom hierarchical exception type for hoyo.

Constructors

ConfigException [Text]

CommandException CommandException

IOException IOError

FileSystemException FileSystemException

ParseException [Text]

MultipleExceptions (NonEmpty HoyoException)

data HoyoMonad a

HoyoMonad is the main monad stack for the hoyo program. It's essentially a wrapper around ExceptT T.Text (ReaderT Env IO): in other words, HoyoMonad a is equivalent to Env -> IO (Either T.Text a)

modifyBookmarks :: ([Bookmark] -> [Bookmark]) -> HoyoMonad ()

Helper function whenever we need to modify the saved bookmarks.

modifyBookmarks f retrieves the current bookmarks, applies f, and saves them back to file.

modifyBookmarksM :: ([Bookmark] -> HoyoMonad [Bookmark]) -> HoyoMonad ()

Helper function twhenever we need to modify the saved bookmarks, and need access to the hoyo environment.

modifyBookmarks f retrieves the current bookmarks, applies f in the hoyo environment, and saves them back to file.

printStderr :: MonadIO m => Text -> m ()

Print to stderr.

printStdout :: MonadIO m => Text -> m ()

Print to stdout.

readInt :: MonadError HoyoException m => Text -> m Int

Try to read an Int.

readBool :: MonadError HoyoException m => Text -> m Bool

Try to read a Bool.

backupFile :: (MonadIO m, MonadError HoyoException m) => FilePath -> String -> m ()

Try to back-up a file. Used when the "backup_before_clear" option is set.

assert :: HoyoException -> HoyoMonad Bool -> HoyoMonad ()

Throw an error if a check fails.

assertVerbose :: HoyoException -> HoyoMonad Bool -> HoyoMonad Bool

Throw an error if a check fails AND the "fail_on_error" flag is set.

Misc

versionString :: String

The current hoyo version.