Hoyo.Bookmark

Description

The Bookmark type provides a representation of bookmarks saved and used by the hoyo program. This module exports some utility datatypes and functions used for working with bookmarks.

The Bookmark type

data Bookmark

Bookmark a directory for easy cd. A bookmark remembers the directory, the index, the creation time, and optionally a user-specified nickname for the bookmark.

Constructors

Bookmark

Fields

_bookmarkDirectory :: !FilePath

_bookmarkIndex :: !Int

_bookmarkCreationTime :: !ZonedTime

_bookmarkName :: !(Maybe Text)

newtype Bookmarks

Wrapper for [Bookmark].

Constructors

Bookmarks

Fields

unBookmarks :: [Bookmark]

data BookmarkSearchTerm

Data-type for represting a bookmark search. You can either search by index or by name. Used by the delete and move commands.

Constructors

SearchIndex Int

SearchName Text

formatBookmark :: Bool -> Int -> Int -> Bookmark -> Text

Format a Bookmark. Used for the "list" command and error reporting during other commands.

formatBookmark displayTime numberWidth bm returns a pretty representation of bm, optionally showing the creation time, and padding the index and directory to a certain width.

formatBookmarks :: Bool -> [Bookmark] -> [Text]

Format a list of Bookmarks. Used for the "list" command and error reporting during other commands

formatBookmark displayTime bms returns a pretty representation of bms, optionally showing the creation time, and padding the indices to line up.

data DefaultBookmark

Default bookmarks to save at init. A default bookmark remembers the directory and optionally a user-specified nickname for the bookmark.

Constructors

DefaultBookmark

Fields

_defaultBookmarkDirectory :: !FilePath

_defaultBookmarkName :: !(Maybe Text)

Working with bookmarks

getBookmarks :: HoyoMonad Bookmarks

Get the bookmarks from the currently used bookmark file.

searchBookmarks :: BookmarkSearchTerm -> Bookmarks -> ([Bookmark], [Bookmark])

searchBookmarks searchTerm bookmarks partitions bookmarks into a list of Bookmarks that match the search term and a list of those that do not.

filterBookmarks :: Maybe Text -> Maybe Text -> Bookmark -> Bool

Given a bookmark name and a bookmark directory, test if a bookmark matches those filters.

filterBookmarkByName :: Maybe Text -> Bookmark -> Bool

A predicate used by filterBookmarks - match on the bookmark name. Note that matching is case-sensitive.

filterBookmarkByDirInfix :: Maybe Text -> Bookmark -> Bool

A predicate used by filterBookmarks - match on the bookmark directory.

bookmarksFromDefault :: MonadIO m => [DefaultBookmark] -> m Bookmarks

Convert a list of DefaultBookmarks to a Bookmarks, assiging indices and/ creation times on the fly.

Parsing/encoding bookmarks from/to TOML

bookmarkCodec :: TomlCodec Bookmark

A TomlCodec for encoding and decoding Bookmarks.

bookmarksCodec :: TomlCodec Bookmarks

A TomlCodec for encoding and decoding Bookmarks.

defaultBookmarkCodec :: TomlCodec DefaultBookmark

A TomlCodec for encoding and decoding Bookmarks.

decodeBookmarks :: Text -> Either Text Bookmarks

Decode a Bookmark from a Text.

decodeBookmarksFile :: MonadIO m => FilePath -> m (Either Text Bookmarks)

Decode a Bookmark from a file.

encodeBookmarks :: Bookmarks -> Text

Encode a Bookmark to a Text.

encodeBookmarksFile :: MonadIO m => FilePath -> Bookmarks -> m ()

Encode a Bookmark to a file.