Hoyo.CLI.Parse

Description

Parse CLI arguments.

Parsing CLI arguments and options

options :: ParserInfo Options

A ParserInfo object containing the necessary information for parsing CLI commands and arguments, and displaying useful help text.

parseOptions :: Parser Options

Parse an Options argument, which includes the command to run and any global options.

parseCommand :: Parser Command

Parse a command and the arguments/options for that command from the command-line arguments.

globalOptions :: Parser GlobalOptions

Parse global options: options that can be set on the command line no matter what sub-command is being run.

For a complete list of the global options, see GlobalOptions or run hoyo --help.

overrideOptions :: Parser OverrideOptions

Parse override options: options that override config settings. This can be useful when you want to temporarily enable functionality for one CLI run, but don't want to change it in the config file.

Parsing specific CLI commands

addCommand :: Parser Command

Parse options for the hoyo add command.

moveCommand :: Parser Command

Parse options for the hoyo move command.

listCommand :: Parser Command

Parse options for the hoyo list command.

clearCommand :: Parser Command

Parse options for the hoyo clear command.

deleteCommand :: Parser Command

Parse options for the hoyo delete command.

checkCommand :: Parser Command

Parse options for the hoyo check command.

defaultCommand :: Parser Command

If hoyo is run with no arguments, we run the "default command" if it's set.

Parsing sub-commands for hoyo config

configCommand :: Parser Command

Parse options for the hoyo config command.

configPrintCommand :: Parser ConfigCommand

Parse options for the hoyo config print sub-command.

configResetCommand :: Parser ConfigCommand

Parse options for the hoyo config reset sub-command.

configSetCommand :: Parser ConfigCommand

Parse options for the hoyo config set sub-command.

configAddDefaultCommand :: Parser ConfigCommand

Parse options for the hoyo config add-default sub-command.

Misc/Utility

splitArgs :: Text -> [String]

Split a string into arguments as they would be interpreted on the command line.

Adapted from this StackOverflow comment.