Hoyo.Command

Description

This module defines data-types and runner functions for the hoyo command-line program.

Running CLI commands

runCommand :: Command -> HoyoMonad ()

Run a Command in the hoyo environment.

modifyBookmarks :: ([Bookmark] -> [Bookmark]) -> HoyoMonad ()

Helper function whenever we need to modify the saved bookmarks.

modifyBookmarks f retrieves the current bookmarks, applies f, and saves them back to file.

modifyBookmarksM :: ([Bookmark] -> HoyoMonad [Bookmark]) -> HoyoMonad ()

Helper function twhenever we need to modify the saved bookmarks, and need access to the hoyo environment.

modifyBookmarks f retrieves the current bookmarks, applies f in the hoyo environment, and saves them back to file.

Specific command runners

runAdd :: AddOptions -> HoyoMonad ()

Run the "add" command: add a new bookmark.

runMove :: MoveOptions -> HoyoMonad ()

Run the "move" command: search for a bookmark and cd to it.

runList :: ListOptions -> HoyoMonad ()

Run the "list" command: list all the saved bookmarks.

runClear :: ClearOptions -> HoyoMonad ()

Run the "clear" command: delete all the saved bookmarks.

runDelete :: DeleteOptions -> HoyoMonad ()

Run the "delete" command: search for a bookmark and delete it.

runRefresh :: RefreshOptions -> HoyoMonad ()

Run the "refresh" command: re-index bookmarks.

runConfig :: ConfigCommand -> HoyoMonad ()

Run the "config" command: dispatch on the given sub-command.

runCheck :: CheckOptions -> FilePath -> FilePath -> IO ()

Run the "config check" command: validate the current config and bookmarks files.

runDefaultCommand :: HoyoMonad ()

Run the default command, if it has been specified by the user.

Types

data Options

The final result of parsing the CLI arguments. Contains a command and all information for that command, and any global options that have been set.

Constructors

Options

Fields

optCommand :: Command

optGlobals :: GlobalOptions

data Command

The core data-type for the hoyo CLI. The Command is parsed from the command-line, then runCommand dispatches on the type.

Constructors

Add AddOptions

Move MoveOptions

List ListOptions

Clear ClearOptions

Delete DeleteOptions

Refresh RefreshOptions

ConfigCmd ConfigCommand

Check CheckOptions

DefaultCommand

Help HelpOptions

data AddOptions

Options for the "add" command to be parsed from the command-line.

Constructors

AddOptions

Fields

addDirectory :: FilePath

addName :: Maybe Text

newtype MoveOptions

Options for the "move" command to be parsed from the command-line.

Constructors

MoveOptions

Fields

moveSearch :: BookmarkSearchTerm

data ListOptions

Options for the "list" command to be parsed from the command-line.

Constructors

ListOptions

Fields

listFilterName :: Maybe Text

listFilterDirectoryInfix :: Maybe Text

listJSONOutput :: Bool

data ClearOptions

Options for the "clear" command to be parsed from the command-line.

Constructors

ClearOptions

newtype DeleteOptions

Options for the "delete" command to be parsed from the command-line.

Constructors

DeleteOptions

Fields

deleteSearch :: BookmarkSearchTerm

data RefreshOptions

Options for the "refresh" command to be parsed from the command-line.

Constructors

RefreshOptions

newtype ConfigPrintOptions

Options for the "config print" command to be parsed from the command-line.

Constructors

ConfigPrintOptions

Fields

configPrintJSONOuput :: Bool

data ConfigResetOptions

Options for the "config reset" command to be parsed from the command-line.

Constructors

ConfigResetOptions

data ConfigSetOptions

Options for the "config set" command to be parsed from the command-line.

Constructors

ConfigSetOptions

Fields

setKey :: Text

setValue :: Text

data ConfigAddDefaultOptions

Options for the "config add-default" command to be parsed from the command-line.

Constructors

ConfigAddDefaultOptions

Fields

addDefaultDir :: FilePath

addDefaultName :: Maybe Text

data ConfigCommand

Options for the "config" command to be parsed from the command-line.

Constructors

Print ConfigPrintOptions

Reset ConfigResetOptions

Set ConfigSetOptions

AddDefaultBookmark ConfigAddDefaultOptions

data CheckOptions

Options for the "check" command to be parsed from the command-line.

Constructors

CheckOptions

Fields

checkConfig :: Bool

checkBookmarks :: Bool

data GlobalOptions

CLI options that can be set regardless of which command is run.

Constructors

GlobalOptions

Fields

globalConfigPath :: Maybe FilePath

dataPath :: Maybe FilePath

overrides :: OverrideOptions

defaultGlobalOptions :: GlobalOptions

Default global options. In general this should do nothing.

data OverrideOptions

Config settings that can be overriden using command-line flags.

Constructors

OverrideOptions

Fields

overrideFailOnError :: MaybeOverride

overrideDisplayCreationTime :: MaybeOverride

overrideEnableClearing :: MaybeOverride

overrideEnableReset :: MaybeOverride

overrideBackupBeforeClear :: MaybeOverride

defaultOverrideOptions :: OverrideOptions

The default behaviour is to override nothing.

overrideConfig :: OverrideOptions -> Config -> Config

Apply the override options to a Config.

overrideEnv :: OverrideOptions -> Env -> Env

Apply the override options to an Env.

verifyOverrides :: OverrideOptions -> Maybe Text

Check that there are no conflicting overrides.

combOverride :: Bool -> Bool -> MaybeOverride

Combine a config flag with a command-line flag, checking for conflicts.

data MaybeOverride

Datatype for representing a command-line settings override.

Constructors

OverrideFalse

OverrideTrue

NoOverride

Conflict