Hoyo.Internal.Command

Description

Internals used by the Hoyo.Command module.

Documentation

combOverride :: Bool -> Bool -> MaybeOverride

Combine a config flag with a command-line flag, checking for conflicts.

overrideFunc :: MaybeOverride -> Bool -> Bool

Convert a MaybeOverride to a function on Bool.

overrideConfig :: OverrideOptions -> Config -> Config

Apply the override options to a Config.

overrideEnv :: OverrideOptions -> Env -> Env

Apply the override options to an Env.

verifyOverrides :: OverrideOptions -> Maybe Text

Check that there are no conflicting overrides.

defaultOverrideOptions :: OverrideOptions

The default behaviour is to override nothing.

defaultGlobalOptions :: GlobalOptions

Default global options. In general this should do nothing.

modifyBookmarks :: ([Bookmark] -> [Bookmark]) -> HoyoMonad ()

Helper function whenever we need to modify the saved bookmarks.

modifyBookmarks f retrieves the current bookmarks, applies f, and saves them back to file.

modifyBookmarksM :: ([Bookmark] -> HoyoMonad [Bookmark]) -> HoyoMonad ()

Helper function twhenever we need to modify the saved bookmarks, and need access to the hoyo environment.

modifyBookmarks f retrieves the current bookmarks, applies f in the hoyo environment, and saves them back to file.

normaliseAndVerifyDirectory :: FilePath -> HoyoMonad FilePath

Normalise a filepath and make sure it's a valid directory.

verifyName :: Text -> HoyoMonad ()

Take a name and make sure it's valid.

testNameUnique :: [Bookmark] -> Text -> HoyoMonad Bool

Given the existing bookmarks and a potential bookmark name, test if the new bookmark will have a unique name.

runAdd :: AddOptions -> HoyoMonad ()

Run the "add" command: add a new bookmark.

runMove :: MoveOptions -> HoyoMonad ()

Run the "move" command: search for a bookmark and cd to it.

runList :: ListOptions -> HoyoMonad ()

Run the "list" command: list all the saved bookmarks.

clearDisabledErrMsg :: Text

Help text displayed when the user tries to run "hoyo clear" when "enable_clear" is set to false.

resetDisabledErrMsg :: Text

Help text displayed when the user tries to run "hoyo config reset" when "enable_reset" is set to false.

getConfirmation :: String -> IO Bool

Prompt the user for confirmation with the given string.

runClear :: ClearOptions -> HoyoMonad ()

Run the "clear" command: delete all the saved bookmarks.

runDelete :: DeleteOptions -> HoyoMonad ()

Run the "delete" command: search for a bookmark and delete it.

runRefresh :: RefreshOptions -> HoyoMonad ()

Run the "refresh" command: re-index bookmarks.

runConfigPrint :: ConfigPrintOptions -> HoyoMonad ()

Run the "config print" command: print the current config.

runConfigReset :: ConfigResetOptions -> HoyoMonad ()

Run the "config reset" command: reset the config to defaultConfig.

runConfigSet :: ConfigSetOptions -> HoyoMonad ()

Run the "config set" command: try to set a key-value pair in the config.

runAddDefault :: ConfigAddDefaultOptions -> HoyoMonad ()

Run the "config add-default" command: try to set a key-value pair in the config.

runConfig :: ConfigCommand -> HoyoMonad ()

Run the "config" command: dispatch on the given sub-command.

runCheckConfig :: FilePath -> IO ()

Check that the config file is valid.

runCheckBookmarks :: FilePath -> IO ()

Check that the bookmarks file is valid.

runCheck :: CheckOptions -> FilePath -> FilePath -> IO ()

Run the "config check" command: validate the current config and bookmarks files.

runDefaultCommand :: HoyoMonad ()

Run the default command, if it has been specified by the user.

runHelp :: HelpOptions -> HoyoMonad ()

Run the "help" command: get help on a specific subcommand.

runCommand :: Command -> HoyoMonad ()

Run a Command in the hoyo environment.