Hoyo.Internal.Config

Description

Internals used by the Hoyo.Config module.

Documentation

configCodec :: TomlCodec Config

A TOML codec describing how to convert a Config to and from its TOML representation.

commandCodec :: Key -> TomlCodec Command

Toml codec using optparse-applicative to parse a default command. Incurs a cyclic dependency which is resolved using Parse.hs-boot.

defaultConfig :: Config

The default config for hoyo.

decodeConfig :: Text -> Either HoyoException Config

Decode a Config from a Text.

decodeConfigFile :: MonadIO m => FilePath -> m (Either HoyoException Config)

Decode a Config from a file.

encodeConfig :: Config -> Text

Encode a Config to a Text.

encodeConfigFile :: MonadIO m => FilePath -> Config -> m ()

Encode a Config to a file.

getKeyVals :: Config -> [(Text, AnyConfigValue)]

Get TOML key-value pairs from a Config.

setConfig :: MonadError HoyoException m => Text -> Text -> Config -> m Config

Try to set a key-value pair in the config.