Hoyo.Internal.Env

Description

Internals used by the Hoyo.Internal module.

Documentation

writeEnv :: MonadIO m => Env -> m ()

Write an Env to file.

readEnv :: MonadIO m => FilePath -> FilePath -> m (Either Text Env)

Read an Env from a file.

initPath :: MonadIO m => FilePath -> m ()

Given a file path, make sure that its directory exists.

initEnv :: MonadIO m => FilePath -> FilePath -> m ()

Given a filepath for the bookmarks file and a filepath for the config file, initialize the respective TOMLs at those locations.

initBookmarksIfNotExists :: (MonadIO m, MonadError Text m) => Config -> FilePath -> m Bookmarks

If the bookmarks path doesn't exist, try to create it.

Returns the newly created Bookmarks object, or the result of parsing the file if it already existed.

initConfigIfNotExists :: (MonadIO m, MonadError Text m) => FilePath -> m Config

If the config path doesn't exist, try to create it.

Returns the newly created Config object, or the result of parsing the file if it already existed.

initEnvIfNotExists :: (MonadIO m, MonadError Text m) => FilePath -> FilePath -> m Env

If the environment files have not been created yet, do so.

Return the Env object.

getEnv :: MonadIO m => FilePath -> FilePath -> m (Either Text Env)

Retrieve an Env from given bookmark- and config- file locations.

defaultConfigPath :: IO FilePath

The default path for the hoyo config. Usually $HOME/.config/hoyo/config.toml

defaultBookmarksPath :: IO FilePath

The default path for hoyo bookmarks. Usually $HOME/.local/share/hoyo/config.toml