Hoyo.Internal.Types

Description

Types used by all the main Hoyo.* modules.

Documentation

data Env

The main hoyo read-only environment. Contains the current saved bookmarks, the current hoyo configuration, and the file locations for each.

Constructors

Env

Fields

_bookmarks :: !Bookmarks

_bookmarksPath :: !FilePath

_config :: !Config

_configPath :: !FilePath

data Bookmark

Bookmark a directory for easy cd. A bookmark remembers the directory, the index, the creation time, and optionally a user-specified nickname for the bookmark.

Constructors

Bookmark

Fields

_bookmarkDirectory :: !FilePath

_bookmarkIndex :: !Int

_bookmarkCreationTime :: !ZonedTime

_bookmarkName :: !(Maybe Text)

data DefaultBookmark

Default bookmarks to save at init. A default bookmark remembers the directory and optionally a user-specified nickname for the bookmark.

Constructors

DefaultBookmark

Fields

_defaultBookmarkDirectory :: !FilePath

_defaultBookmarkName :: !(Maybe Text)

newtype Bookmarks

Wrapper for [Bookmark].

Constructors

Bookmarks

Fields

unBookmarks :: [Bookmark]

data BookmarkSearchTerm

Data-type for represting a bookmark search. You can either search by index or by name. Used by the delete and move commands.

Constructors

SearchIndex Int

SearchName Text

data ConfigValueType

The types of config values allowed in the Hoyo config.

Constructors

TBool

TDefaultBookmark

TCommand

TList ConfigValueType

TMaybe ConfigValueType

data ConfigValue (t :: ConfigValueType) where

Values in the Hoyo config. Using a GADT parameterised by ConfigValueType gives us stricter type safety.

Constructors

BoolV :: Bool -> ConfigValue 'TBool

DefaultBookmarkV :: DefaultBookmark -> ConfigValue 'TDefaultBookmark

CommandV :: Command -> ConfigValue 'TCommand

ListOfV :: forall (a :: ConfigValueType). [ConfigValue a] -> ConfigValue ('TList a)

MaybeV :: forall (a :: ConfigValueType). Maybe (ConfigValue a) -> ConfigValue ('TMaybe a)

data AnyConfigValue

Existential wrapper around ConfigValue.

Constructors

forall (t :: ConfigValueType). AnyConfigValue (ConfigValue t)

data Config

A representation of hoyo settings.

Constructors

Config

Fields

_failOnError :: !(ConfigValue 'TBool)

_displayCreationTime :: !(ConfigValue 'TBool)

_enableClearing :: !(ConfigValue 'TBool)

_enableReset :: !(ConfigValue 'TBool)

_backupBeforeClear :: !(ConfigValue 'TBool)

_defaultBookmarks :: !(ConfigValue ('TList 'TDefaultBookmark))

_defaultCommand :: !(ConfigValue ('TMaybe 'TCommand))

data SearchException

Report an exception while searching for a bookmark.

Constructors

NothingFound BookmarkSearchTerm

TooManyResults BookmarkSearchTerm [Text]

data CommandException

Report an exception while running a command.

Constructors

SearchException SearchException

InvalidArgumentException [Text]

LoopException

data FileSystemException

Report a file system exception.

Constructors

NoFileException FilePath

NoDirException FilePath

data HoyoException

A custom hierarchical exception type for hoyo.

Constructors

ConfigException [Text]

CommandException CommandException

IOException IOError

FileSystemException FileSystemException

ParseException [Text]

MultipleExceptions (NonEmpty HoyoException)

newtype HoyoMonad a

HoyoMonad is the main monad stack for the hoyo program. It's essentially a wrapper around ExceptT T.Text (ReaderT Env IO): in other words, HoyoMonad a is equivalent to Env -> IO (Either T.Text a)

Constructors

HoyoMonad

Fields

unHoyo :: ExceptT HoyoException (ReaderT Env IO) a

data AddOptions

Options for the "add" command to be parsed from the command-line.

Constructors

AddOptions

Fields

addDirectory :: FilePath

addName :: Maybe Text

newtype MoveOptions

Options for the "move" command to be parsed from the command-line.

Constructors

MoveOptions

Fields

moveSearch :: BookmarkSearchTerm

data ListOptions

Options for the "list" command to be parsed from the command-line.

Constructors

ListOptions

Fields

listFilterName :: Maybe Text

listFilterDirectoryInfix :: Maybe Text

listJSONOutput :: Bool

data ClearOptions

Options for the "clear" command to be parsed from the command-line.

Constructors

ClearOptions

newtype DeleteOptions

Options for the "delete" command to be parsed from the command-line.

Constructors

DeleteOptions

Fields

deleteSearch :: BookmarkSearchTerm

data RefreshOptions

Options for the "refresh" command to be parsed from the command-line.

Constructors

RefreshOptions

newtype ConfigPrintOptions

Options for the "config print" command to be parsed from the command-line.

Constructors

ConfigPrintOptions

Fields

configPrintJSONOuput :: Bool

data ConfigResetOptions

Options for the "config reset" command to be parsed from the command-line.

Constructors

ConfigResetOptions

data ConfigSetOptions

Options for the "config set" command to be parsed from the command-line.

Constructors

ConfigSetOptions

Fields

setKey :: Text

setValue :: Text

data ConfigAddDefaultOptions

Options for the "config add-default" command to be parsed from the command-line.

Constructors

ConfigAddDefaultOptions

Fields

addDefaultDir :: FilePath

addDefaultName :: Maybe Text

data ConfigCommand

Options for the "config" command to be parsed from the command-line.

Constructors

Print ConfigPrintOptions

Reset ConfigResetOptions

Set ConfigSetOptions

AddDefaultBookmark ConfigAddDefaultOptions

data CheckOptions

Options for the "check" command to be parsed from the command-line.

Constructors

CheckOptions

Fields

checkConfig :: Bool

checkBookmarks :: Bool

newtype HelpOptions

Options for the "help" command to be parsed from the command-line.

Constructors

HelpOptions

Fields

helpSubcommand :: Maybe Text

data Command

The core data-type for the hoyo CLI. The Command is parsed from the command-line, then runCommand dispatches on the type.

Constructors

Add AddOptions

Move MoveOptions

List ListOptions

Clear ClearOptions

Delete DeleteOptions

Refresh RefreshOptions

ConfigCmd ConfigCommand

Check CheckOptions

DefaultCommand

Help HelpOptions

data MaybeOverride

Datatype for representing a command-line settings override.

Constructors

OverrideFalse

OverrideTrue

NoOverride

Conflict

data OverrideOptions

Config settings that can be overriden using command-line flags.

Constructors

OverrideOptions

Fields

overrideFailOnError :: MaybeOverride

overrideDisplayCreationTime :: MaybeOverride

overrideEnableClearing :: MaybeOverride

overrideEnableReset :: MaybeOverride

overrideBackupBeforeClear :: MaybeOverride

data GlobalOptions

CLI options that can be set regardless of which command is run.

Constructors

GlobalOptions

Fields

globalConfigPath :: Maybe FilePath

dataPath :: Maybe FilePath

overrides :: OverrideOptions

data Options

The final result of parsing the CLI arguments. Contains a command and all information for that command, and any global options that have been set.

Constructors

Options

Fields

optCommand :: Command

optGlobals :: GlobalOptions