Hoyo.Internal.Utils

Description

Utility functions used by all the main Hoyo.* modules.

Documentation

asks' :: MonadReader a m => SimpleGetter a b -> m b

A version of the lens "use" function for MonadReader.

maximumDefault :: Ord a => a -> [a] -> a

Take the maximum of a list, with a default value if the list is empty.

assert :: Text -> HoyoMonad Bool -> HoyoMonad ()

Throw an error if a check fails.

assertVerbose :: Text -> HoyoMonad Bool -> HoyoMonad Bool

Throw an error if a check fails AND the "fail_on_error" flag is set.

getBackupFile :: (MonadIO m, MonadError Text m) => FilePath -> String -> m FilePath

Given a file name and an extension, try to find a suitable path for backing up that file. Used by backupFile.

backupFile :: (MonadIO m, MonadError Text m) => FilePath -> String -> m ()

Try to back-up a file. Used when the "backup_before_clear" option is set.

readBool :: MonadError Text m => Text -> m Bool

Try to read a Bool.

readInt :: MonadError Text m => Text -> m Int

Try to read an Int.

printStderr :: MonadIO m => Text -> m ()

Print to stderr.

printStdout :: MonadIO m => Text -> m ()

Print to stdout.

pageLines :: MonadIO m => [Text] -> m ()

Print or page lines.

formatBookmark :: Bool -> Int -> Int -> Bookmark -> Text

Format a Bookmark. Used for the "list" command and error reporting during other commands.

formatBookmark displayTime numberWidth bm returns a pretty representation of bm, optionally showing the creation time, and padding the index and directory to a certain width.

formatBookmarks :: Bool -> [Bookmark] -> [Text]

Format a list of Bookmarks. Used for the "list" command and error reporting during other commands

formatBookmark displayTime bms returns a pretty representation of bms, optionally showing the creation time, and padding the indices to line up.

formatDefaultBookmark :: DefaultBookmark -> Text

Format a DefaultBookmark. Used for the "config print" command and error reporting during other commands.

formatDefaultBookmark bm returns a pretty representation of bm.

tshow :: Show a => a -> Text

Show a value as a Text instead of a String.

formatConfigValue :: AnyConfigValue -> Text

Format a config value. Used for the "config print" command.