HoYo.Internal.Command

Description

Internals used by the HoYo.Command module.

Documentation

combOverride :: Bool -> Bool -> MaybeOverride

Combine a config flag with a command-line flag, checking for conflicts.

overrideFunc :: MaybeOverride -> Bool -> Bool

Convert a MaybeOverride to a function on Bool.

overrideConfig :: OverrideOptions -> Config -> Config

Apply the override options to a Config.

overrideEnv :: OverrideOptions -> Env -> Env

Apply the override options to an Env.

verifyOverrides :: OverrideOptions -> Maybe Text

Check that there are no conflicting overrides.

defaultOverrideOptions :: OverrideOptions

The default behaviour is to override nothing.

defaultGlobalOptions :: GlobalOptions

Default global options. In general this should do nothing.

modifyBookmarks :: ([Bookmark] -> [Bookmark]) -> HoYoMonad ()

Helper function whenever we need to modify the saved bookmarks.

modifyBookmarks f retrieves the current bookmarks, applies f, and saves them back to file.

modifyBookmarksM :: ([Bookmark] -> HoYoMonad [Bookmark]) -> HoYoMonad ()

Helper function twhenever we need to modify the saved bookmarks, and need access to the hoyo environment.

modifyBookmarks f retrieves the current bookmarks, applies f in the hoyo environment, and saves them back to file.

normaliseAndVerifyDirectory :: TFilePath -> HoYoMonad TFilePath

Normalise a filepath and make sure it's a valid directory.

verifyName :: Text -> HoYoMonad ()

Take a name and make sure it's valid.

testNameUnique :: [Bookmark] -> Text -> HoYoMonad Bool

Given the existing bookmarks and a potential bookmark name, test if the new bookmark will have a unique name.

runAdd :: AddOptions -> HoYoMonad ()

Run the "add" command: add a new bookmark.

runMove :: MoveOptions -> HoYoMonad ()

Run the "move" command: search for a bookmark and cd to it.

runList :: ListOptions -> HoYoMonad ()

Run the "list" command: list all the saved bookmarks.

clearDisabledErrMsg :: Text

Help text displayed when the user tries to run "hoyo clear" when "enable_clear" is set to false.

resetDisabledErrMsg :: Text

Help text displayed when the user tries to run "hoyo config reset" when "enable_reset" is set to false.

runClear :: ClearOptions -> HoYoMonad ()

Run the "clear" command: delete all the saved bookmarks.

runDelete :: DeleteOptions -> HoYoMonad ()

Run the "delete" command: search for a bookmark and delete it.

runRefresh :: RefreshOptions -> HoYoMonad ()

Run the "refresh" command: re-index bookmarks.

runConfigPrint :: ConfigPrintOptions -> HoYoMonad ()

Run the "config print" command: print the current config.

runConfigReset :: ConfigResetOptions -> HoYoMonad ()

Run the "config reset" command: reset the config to defaultConfig.

runConfigSet :: ConfigSetOptions -> HoYoMonad ()

Run the "config set" command: try to set a key-value pair in the config.

runAddDefault :: ConfigAddDefaultOptions -> HoYoMonad ()

Run the "config add-default" command: try to set a key-value pair in the config.

runConfig :: ConfigCommand -> HoYoMonad ()

Run the "config" command: dispatch on the given sub-command.

runCheckConfig :: TFilePath -> IO ()

Check that the config file is valid.

runCheckBookmarks :: TFilePath -> IO ()

Check that the bookmarks file is valid.

runCheck :: CheckOptions -> TFilePath -> TFilePath -> IO ()

Run the "config check" command: validate the current config and bookmarks files.

runDefaultCommand :: HoYoMonad ()

Run the default command, if it has been specified by the user.

runCommand :: Command -> HoYoMonad ()

Run a Command in the hoyo environment.