HoYo.Internal.Types

Description

Types used by all the main HoYo.* modules.

Documentation

type TFilePath = Text

A Text version of FilePath.

data Env

The main hoyo read-only environment. Contains the current saved bookmarks, the current hoyo configuration, and the file locations for each.

Constructors

Env

Fields

_bookmarks :: !Bookmarks

_bookmarksPath :: !TFilePath

_config :: !Config

_configPath :: !TFilePath

data Bookmark

Bookmark a directory for easy cd. A bookmark remembers the directory, the index, the creation time, and optionally a user-specified nickname for the bookmark.

Constructors

Bookmark

Fields

_bookmarkDirectory :: !TFilePath

_bookmarkIndex :: !Int

_bookmarkCreationTime :: !ZonedTime

_bookmarkName :: !(Maybe Text)

data DefaultBookmark

Default bookmarks to save at init. A default bookmark remembers the directory and optionally a user-specified nickname for the bookmark.

Constructors

DefaultBookmark

Fields

_defaultBookmarkDirectory :: !TFilePath

_defaultBookmarkName :: !(Maybe Text)

newtype Bookmarks

Wrapper for [Bookmark].

Constructors

Bookmarks

Fields

unBookmarks :: [Bookmark]

data BookmarkSearchTerm

Data-type for represting a bookmark search. You can either search by index or by name. Used by the delete and move commands.

Constructors

SearchIndex Int

SearchName Text

data ConfigValueType

The types of config values allowed in the HoYo config.

Constructors

TBool

TDefaultBookmark

TCommand

TList ConfigValueType

TMaybe ConfigValueType

data ConfigValue (t :: ConfigValueType) where

Values in the HoYo config. Using a GADT parameterised by ConfigValueType gives us stricter type safety.

Constructors

BoolV :: Bool -> ConfigValue 'TBool

DefaultBookmarkV :: DefaultBookmark -> ConfigValue 'TDefaultBookmark

CommandV :: Command -> ConfigValue 'TCommand

ListOfV :: forall (a :: ConfigValueType). [ConfigValue a] -> ConfigValue ('TList a)

MaybeV :: forall (a :: ConfigValueType). Maybe (ConfigValue a) -> ConfigValue ('TMaybe a)

data AnyConfigValue

Existential wrapper around ConfigValue.

Constructors

forall (t :: ConfigValueType). AnyConfigValue (ConfigValue t)

data Config

A representation of hoyo settings.

Constructors

Config

Fields

_failOnError :: !(ConfigValue 'TBool)

_displayCreationTime :: !(ConfigValue 'TBool)

_enableClearing :: !(ConfigValue 'TBool)

_enableReset :: !(ConfigValue 'TBool)

_backupBeforeClear :: !(ConfigValue 'TBool)

_defaultBookmarks :: !(ConfigValue ('TList 'TDefaultBookmark))

_defaultCommand :: !(ConfigValue ('TMaybe 'TCommand))

newtype HoYoMonad a

HoYoMonad is the main monad stack for the hoyo program. It's essentially a wrapper around ExceptT T.Text (ReaderT Env IO): in other words, HoYoMonad a is equivalent to Env -> IO (Either T.Text a)

Constructors

HoYoMonad

Fields

unHoYo :: ExceptT Text (ReaderT Env IO) a

data AddOptions

Options for the "add" command to be parsed from the command-line.

Constructors

AddOptions

Fields

addDirectory :: TFilePath

addName :: Maybe Text

newtype MoveOptions

Options for the "move" command to be parsed from the command-line.

Constructors

MoveOptions

Fields

moveSearch :: BookmarkSearchTerm

data ListOptions

Options for the "list" command to be parsed from the command-line.

Constructors

ListOptions

Fields

listFilterName :: Maybe Text

listFilterDirectoryInfix :: Maybe Text

data ClearOptions

Options for the "clear" command to be parsed from the command-line.

Constructors

ClearOptions

newtype DeleteOptions

Options for the "delete" command to be parsed from the command-line.

Constructors

DeleteOptions

Fields

deleteSearch :: BookmarkSearchTerm

data RefreshOptions

Options for the "refresh" command to be parsed from the command-line.

Constructors

RefreshOptions

data ConfigPrintOptions

Options for the "config print" command to be parsed from the command-line.

Constructors

ConfigPrintOptions

data ConfigResetOptions

Options for the "config reset" command to be parsed from the command-line.

Constructors

ConfigResetOptions

data ConfigSetOptions

Options for the "config set" command to be parsed from the command-line.

Constructors

ConfigSetOptions

Fields

setKey :: Text

setValue :: Text

data ConfigAddDefaultOptions

Options for the "config add-default" command to be parsed from the command-line.

Constructors

ConfigAddDefaultOptions

Fields

addDefaultDir :: TFilePath

addDefaultName :: Maybe Text

data ConfigCommand

Options for the "config" command to be parsed from the command-line.

Constructors

Print ConfigPrintOptions

Reset ConfigResetOptions

Set ConfigSetOptions

AddDefaultBookmark ConfigAddDefaultOptions

data CheckOptions

Options for the "check" command to be parsed from the command-line.

Constructors

CheckOptions

Fields

checkConfig :: Bool

checkBookmarks :: Bool

data Command

The core data-type for the hoyo CLI. The Command is parsed from the command-line, then runCommand dispatches on the type.

Constructors

Add AddOptions

Move MoveOptions

List ListOptions

Clear ClearOptions

Delete DeleteOptions

Refresh RefreshOptions

ConfigCmd ConfigCommand

Check CheckOptions

DefaultCommand

data MaybeOverride

Datatype for representing a command-line settings override.

Constructors

OverrideFalse

OverrideTrue

NoOverride

Conflict

data OverrideOptions

Config settings that can be overriden using command-line flags.

Constructors

OverrideOptions

Fields

overrideFailOnError :: MaybeOverride

overrideDisplayCreationTime :: MaybeOverride

overrideEnableClearing :: MaybeOverride

overrideEnableReset :: MaybeOverride

overrideBackupBeforeClear :: MaybeOverride

data GlobalOptions

CLI options that can be set regardless of which command is run.

Constructors

GlobalOptions

Fields

globalConfigPath :: Maybe TFilePath

dataPath :: Maybe TFilePath

overrides :: OverrideOptions

data Options

The final result of parsing the CLI arguments. Contains a command and all information for that command, and any global options that have been set.

Constructors

Options

Fields

optCommand :: Command

optGlobals :: GlobalOptions